Externalized the logic of processing pie data.
